pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ravvi-poker"
dynamic = ["version"]
authors = [
    {name = "Alexander Keda", email = "alexander.keda@ravvi.net"},
    {name = "Stanislav Chuprinskiy", email = "stanislav.chuprinskiy@ravvi.net"}
]
description = "Ravvi Poker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg",
    "psycopg_pool",
    "fastapi",
    "python-multipart",
    "passlib",
    "PyJWT",
    "PyYAML",
    "uvicorn",
    "gunicorn",
    "websockets",
    "pydantic[email]",
    "Pillow",
    "python-magic",
    "aiohttp"
]

[project.optional-dependencies]
tests = [
    "coverage",
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "pytest-dependency",
    "httpx"
]

[project.scripts]
ravvi_poker_db = "ravvi_poker.db.cli:main"
ravvi_poker_api = "ravvi_poker.api.cli:main"
#ravvi_poker_ws = "ravvi_poker.ws.cli:main"
ravvi_poker_engine = "ravvi_poker.engine.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ravvi_poker*"]
namespaces = false

[tool.setuptools.package-data]
"ravvi_poker.db.schema" = ["*.sql"]
"ravvi_poker.db.deploy" = ["*.sql"]

[tool.setuptools.dynamic]
version = {attr = "ravvi_poker.build.version"}

[tool.black]
line-length = 120